Updated documentation
Ported API reference examples into class and method javadoc.
Fixed all javadoc warnings.
Removed javadoc that provided no more information than the method
signature.
Added overview and package info.
Added javadoc overview to replace large parts of README.md.
Created package-info.java files.
Updated build.grade to include http classes in javadoc and the overview.
Updated README and CHANGES.
Updated javadoc for http package.
Added hyperlink to github page.
Added compatibility information.
Updated external links to open outside of the frames.
Changed instances of holds or represents to encapsulates.
Ensured escaping of “ in selector javadoc.

BugId: 47708
