+++ CHANGES.md
@@ -1,4 +1,5 @@
 # Unreleased
+- [NEW] `DesignDocument.MapReduce` now has a setter for the `dbcopy` field.
 - [NEW] Requests for the `_all_docs` endpoint are made via `Database#getAllDocsRequestBuilder()`
   instead of using a view.
 - [NEW] Introduced new view query API. More information is available in the javadoc,
@@ -7,9 +8,12 @@
   For example, if you were using the `InputStream` directly for streaming API parsing with an alternative
   JSON library we might be able to make this easier by handling the streams and providing a callback.
 - [NEW] Optional OkHttp dependency for per CloudantClient instance connection pooling.
+- [BREAKING CHANGE] Removed `Database.batch(Object)` method.
 - [BREAKING CHANGE] JVM `http.maxConnections` configured pool is used by default for connection pooling.
 - [BREAKING CHANGE] Removed Apache HttpClient dependency. API methods that used HttpClient classes
   (e.g. `executeRequest`) now use `HttpConnection` instead.
+- [BREAKING CHANGE] `CloudantClient` public constructors and `ConnectionOptions` have been removed.
+  `'CloudantClient` instances are now created and have options configured using `ClientBuilder`.
 - [BREAKING CHANGE] Removed version 1.x view query API.
 - [BREAKING CHANGE] LightCouch classes moved to package com.cloudant.client.org.lightcouch.
   This should only have a visible impact for `CouchDbException` and its subclasses.
@@ -19,6 +23,8 @@
   `DesignDocumentManager.fromDirectory(File)`.
   More information is available in the javadoc, including usage for de-serializing design document
   javascript files to DesignDocument objects.
+- [FIX] Use the default port for the protocol when a client instance is created from a URL without
+  specifying a port.
 
 # 1.2.3 (2015-10-14)
 - [NEW] Added Basic Auth for HTTP proxies. Configure via `ConnectOption#setProxyUser`
