+++ src/main/java/com/cloudant/client/api/views/ViewMultipleRequest.java
@@ -74,7 +74,7 @@ public interface ViewMultipleRequest<K, V> {
      * Perform a single POST request to get the responses for the queries built into this request.
      *
      * @return a list of the responses, one for each query
-     * @throws IOException
+     * @throws IOException if there is an error communicating with the server
      * @since 2.0.0
      */
     List<ViewResponse<K, V>> getViewResponses() throws IOException;
