+++ src/main/java/com/cloudant/http/HttpConnectionInterceptor.java
@@ -1,4 +1,7 @@
 package com.cloudant.http;
 
+/**
+ * A marker interface for all HttpConnectionInterceptors.
+ */
 public interface HttpConnectionInterceptor {
 }
