+++ src/main/java/com/cloudant/http/HttpConnectionResponseInterceptor.java
@@ -15,17 +15,19 @@
 package com.cloudant.http;
 
 /**
- A Response Interceptor is run after the response is obtained from the
- server but before the output stream is returned to the original client. The Response
- Interceptor enables two main behaviours:
-<ul>
-  <li>Modifying the response for every request</li>
-
- <li> Replaying a (potentially modified) request by reacting to the
- response.</li>
- </ul>
-
- Interceptors are executed in a pipeline and modify the context in a serial fashion.
+ * A Response Interceptor is run after the response is obtained from the
+ * server but before the output stream is returned to the original client. The Response
+ * Interceptor enables two main behaviours:
+ * <ul>
+ * <li>Modifying the response for every request</li>
+ *
+ * <li> Replaying a (potentially modified) request by reacting to the
+ * response.</li>
+ * </ul>
+ *
+ * Interceptors are executed in a pipeline and modify the context in a serial fashion.
+ *
+ * @since 2.0.0
  */
 public interface HttpConnectionResponseInterceptor extends HttpConnectionInterceptor {
 
@@ -34,9 +36,10 @@ public interface HttpConnectionResponseInterceptor extends HttpConnectionInterce
      *
      * This method <strong>must not</strong> do any of the following
      * <ul>
-     *     <li>Return null</li>
-     *     <li>Read the response stream</li>
+     * <li>Return null</li>
+     * <li>Read the response stream</li>
      * </ul>
+     *
      * @param context Input context
      * @return Output context
      */
