+++ src/main/java/com/cloudant/http/internal/Base64OutputStreamFactory.java
@@ -12,7 +12,7 @@
  * and limitations under the License.
  */
 
-package com.cloudant.http;
+package com.cloudant.http.internal;
 
 import java.io.OutputStream;
 import java.lang.reflect.Constructor;
